로그인한 작성자가 생성한 글, 댓글만 수정,삭제 가능

+++ src/main/kotlin/com/teamsparta/todoproject/domain/comment/service/CommentServiceImpl.kt
@@ -8,9 +8,11 @@ import com.teamsparta.todoproject.domain.comment.model.Comment
 import com.teamsparta.todoproject.domain.comment.model.toResponse
 import com.teamsparta.todoproject.domain.comment.repository.CommentRepository
 import com.teamsparta.todoproject.domain.exception.PostNotFoundException
+import com.teamsparta.todoproject.domain.exception.dto.NotAuthorizationException
 import com.teamsparta.todoproject.domain.post.repository.PostRepository
 import com.teamsparta.todoproject.domain.user.exception.UserNotFoundException
 import com.teamsparta.todoproject.domain.user.repository.UserRepository
+import com.teamsparta.todoproject.infra.security.UserPrincipal
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -28,9 +30,9 @@ class CommentServiceImpl(
     }
 
     @Transactional
-    override fun createComment(postId: Long, request: CreateCommentRequest): CommentResponse {
+    override fun createComment(userPrincipal: UserPrincipal, postId: Long, request: CreateCommentRequest): CommentResponse {
         val post = postRepository.findByIdOrNull(postId)?: throw CommentNotFoundException("Post", postId)
-        val user = userRepository.findByIdOrNull(request.userId)?: throw UserNotFoundException("User", null)
+        val user = userRepository.findByIdOrNull(userPrincipal.id)?: throw UserNotFoundException("User", null)
         val comment = Comment(
             description = request.description,
             post = post,
@@ -42,8 +44,10 @@ class CommentServiceImpl(
     }
 
     @Transactional
-    override fun updateComment(postId: Long, commentId: Long, request: UpdateCommentRequest): CommentResponse {
+    override fun updateComment(userPrincipal: UserPrincipal, postId: Long, commentId: Long, request: UpdateCommentRequest): CommentResponse {
         val comment = commentRepository.findByPostIdAndId(postId, commentId)?: throw CommentNotFoundException("Comment", commentId)
+        userRepository.findByIdOrNull(userPrincipal.id)?: throw UserNotFoundException("User", null)
+        if(comment.user.id != userPrincipal.id) throw NotAuthorizationException()
         val (description) = request
 
         comment.description = description
@@ -52,9 +56,11 @@ class CommentServiceImpl(
     }
 
     @Transactional
-    override fun deleteComment(postId: Long, commentId: Long) {
+    override fun deleteComment(userPrincipal: UserPrincipal, postId: Long, commentId: Long) {
         val post = postRepository.findByIdOrNull(postId)?: throw PostNotFoundException("Post", postId)
         val comment = commentRepository.findByIdOrNull(commentId)?: throw CommentNotFoundException("Comment", commentId)
+        userRepository.findByIdOrNull(userPrincipal.id)?: throw UserNotFoundException("User", null)
+        if(comment.user.id != userPrincipal.id) throw NotAuthorizationException()
         post.deleteComment(comment)
         postRepository.save(post)
     }
